+++ packages/ai/CHANGELOG.md
@@ -1,5 +1,19 @@
 # ai
 
+## 7.0.0-canary.152
+
+### Patch Changes
+
+- Updated dependencies [d4d4a5e]
+  - @ai-sdk/gateway@4.0.0-canary.91
+
+## 7.0.0-canary.151
+
+### Patch Changes
+
+- Updated dependencies [8b811d8]
+  - @ai-sdk/gateway@4.0.0-canary.90
+
 ## 7.0.0-canary.150
 
 ### Patch Changes