+++ packages/react/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @ai-sdk/react
 
+## 4.0.0-canary.154
+
+### Patch Changes
+
+- ai@7.0.0-canary.152
+
+## 4.0.0-canary.153
+
+### Patch Changes
+
+- ai@7.0.0-canary.151
+
 ## 4.0.0-canary.152
 
 ### Patch Changes