+++ packages/rsc/tests/e2e/next-server/CHANGELOG.md
@@ -4,6 +4,18 @@
 
 ### Patch Changes
 
+- ai@7.0.0-canary.152
+
+## 0.0.1-canary.0
+
+### Patch Changes
+
+- ai@7.0.0-canary.151
+
+## 0.0.1-canary.0
+
+### Patch Changes
+
 - ai@7.0.0-canary.150
 
 ## 0.0.1-canary.0